AP-8036: Change type of parameter from int to long in zeroPad method

+++ org.knime.dl.python/src/org/knime/dl/python/core/data/DLPythonAbstractDataBuffer.java
@@ -114,7 +114,7 @@ public abstract class DLPythonAbstractDataBuffer<B extends DLWrappingDataBuffer<
 	}
 
 	@Override
-	public void zeroPad(final int length) throws IllegalArgumentException, BufferOverflowException {
+	public void zeroPad(final long length) throws IllegalArgumentException, BufferOverflowException {
 		m_buffer.zeroPad(length);
 	}
