+++ org.knime.dl/src/org/knime/dl/core/data/DLDefaultFloatBuffer.java
@@ -161,7 +161,7 @@ public class DLDefaultFloatBuffer extends DLAbstractWrappingDataBuffer<float[]>
 	}
 
 	@Override
-	public void zeroPad(final int length) throws IllegalArgumentException, BufferOverflowException {
+	public void zeroPad(final long length) throws IllegalArgumentException, BufferOverflowException {
 		checkArgument(length > 0);
 		checkOverflow(m_nextWrite + length <= m_capacity);
 		for (int i = 0; i < length; i++) {
